revert: roll shields PR files back to 0c3b6a4fa state
