+++ src/lib/shields/index.ts
@@ -16,6 +16,10 @@ const { fork } = require("child_process");
 const { randomBytes } = require("crypto");
 const { run, runCapture, validateName } = require("../runner");
 const { dockerExecFileSync } = require("../adapters/docker/exec");
+const { dockerCapture } = require("../adapters/docker/run");
+const registry = require("../state/registry") as {
+  getSandbox?: (name: string) => { openshellDriver?: string | null } | null;
+};
 const {
   buildPolicyGetCommand,
   buildPolicySetCommand,
@@ -39,9 +43,6 @@ const { resolveNemoclawStateDir } = require("../state/paths");
 const { appendAuditEntry } = require("./audit");
 const { resolveAgentConfig } = require("../sandbox/config");
 const {
-  privilegedSandboxExecArgv,
-}: typeof import("../sandbox/privileged-exec") = require("../sandbox/privileged-exec");
-const {
   buildRuntimePermissivePolicy,
 }: typeof import("./permissive-runtime") = require("./permissive-runtime");
 const { cleanupTempDir } = require("../onboard/temp-files");
@@ -55,11 +56,65 @@ const STATE_DIR = resolveNemoclawStateDir();
 // privileged sandbox exec — bypasses the sandbox's Landlock context
 //
 // openshell sandbox exec runs commands INSIDE the Landlock domain, so it
-// can't modify read_only paths or change chattr flags. Privileged Docker exec
-// starts a new root process that does NOT inherit the Landlock ruleset.
-// OpenShell gateways expose sandboxes as Docker containers, so we exec into
-// the sandbox container directly as root.
+// can't modify read_only paths or change chattr flags. kubectl exec starts
+// a new process in the pod that does NOT inherit the Landlock ruleset.
+// On the legacy gateway we reach kubectl via the K3s container. On the
+// Docker-driver gateway there is no K3s container, so we exec into the
+// sandbox Docker container directly as root.
 // ---------------------------------------------------------------------------
+
+const K3S_CONTAINER = "openshell-cluster-nemoclaw";
+
+function resolveDockerDriverSandboxContainer(
+  sandboxName: string,
+): string | null {
+  try {
+    if (registry.getSandbox?.(sandboxName)?.openshellDriver !== "docker") {
+      return null;
+    }
+  } catch {
+    return null;
+  }
+  const prefix = `openshell-${sandboxName}-`;
+  const exact = `openshell-${sandboxName}`;
+  const output = dockerCapture(["ps", "--format", "{{.Names}}"], {
+    ignoreError: true,
+  });
+  return (
+    output
+      .split("\n")
+      .map((line: string) => line.trim())
+      .find((name: string) => name === exact || name.startsWith(prefix)) || null
+  );
+}
+
+function kubectlExecArgv(sandboxName: string, cmd: string[]): string[] {
+  return [
+    "exec",
+    K3S_CONTAINER,
+    "kubectl",
+    "exec",
+    "-n",
+    "openshell",
+    sandboxName,
+    "-c",
+    "agent",
+    "--",
+    ...cmd,
+  ];
+}
+
+function privilegedSandboxExecArgv(
+  sandboxName: string,
+  cmd: string[],
+): string[] {
+  const dockerDriverContainer =
+    resolveDockerDriverSandboxContainer(sandboxName);
+  if (dockerDriverContainer) {
+    return ["exec", "--user", "root", dockerDriverContainer, ...cmd];
+  }
+  return kubectlExecArgv(sandboxName, cmd);
+}
 
 function privilegedSandboxExec(sandboxName: string, cmd: string[]): void {
   dockerExecFileSync(privilegedSandboxExecArgv(sandboxName, cmd), {
@@ -448,7 +503,7 @@ function assertNoLegacyStateLayout(
 // user and the gateway UID can write the mutable config tree. Hermes keeps its
 // tighter single-user layout.
 //
-// Note on chattr: best-effort — it may silently fail if privileged exec
+// Note on chattr: best-effort — it may silently fail if kubectl exec
 // lacks CAP_LINUX_IMMUTABLE or if the file was never immutable. That's fine:
 // the file becomes writable through the permissive policy (disables Landlock
 // read_only) + chown/chmod below.
@@ -582,7 +637,7 @@ function unlockAgentConfig(
 //   2. UNIX permissions — 444 root:root (mandatory, verified here)
 //   3. chattr +i immutable bit — defense-in-depth (best-effort)
 //
-// Layer 3 is best-effort because privileged exec may lack
+// Layer 3 is best-effort because kubectl exec may lack
 // CAP_LINUX_IMMUTABLE. Layers 1+2 are sufficient. We still attempt it
 // in case the runtime environment supports it.
 // ---------------------------------------------------------------------------
@@ -623,7 +678,7 @@ function lockAgentConfig(
     errors.push("chown root:root config dir");
   }
 
-  // Best-effort: privileged exec may lack CAP_LINUX_IMMUTABLE. Track the
+  // Best-effort: kubectl exec may lack CAP_LINUX_IMMUTABLE. Track the
   // result so verification doesn't require something that was never there.
   let chattrSucceeded = true;
   for (const f of filesToLock) {
@@ -705,7 +760,7 @@ function rollbackShieldsDown(
   } else {
     console.error("  Config remains unlocked — manual intervention required.");
     console.error(
-      `  Restore sandbox access, then run: nemoclaw ${sandboxName} shields up`,
+      `  Re-lock manually via kubectl exec, then run: nemoclaw ${sandboxName} shields up`,
     );
   }
 }
@@ -968,7 +1023,7 @@ function shieldsDown(sandboxName: string, opts: ShieldsDownOpts = {}): void {
 
   // 4. Start auto-restore timer (detached child process), unless skipped.
   //    Pass the absolute restore time, not a relative timeout. Steps 1-2b
-  //    can take minutes (policy apply + privileged chmod), so a relative timeout
+  //    can take minutes (policy apply + kubectl chmod), so a relative timeout
   //    passed at fork time would fire too early.
   if (!opts.skipTimer) {
     const restoreAt = new Date(Date.now() + timeoutSeconds * 1000);
@@ -1098,13 +1153,13 @@ function shieldsUp(sandboxName: string, opts: { throwOnError?: boolean } = {}): 
         "  Config remains unlocked — manual intervention required.",
       );
       console.error(
-        `  Restore sandbox access, then run: nemoclaw ${sandboxName} shields up`,
+        `  Re-lock manually via kubectl exec, then run: nemoclaw ${sandboxName} shields up`,
       );
       return failShieldsCommand(activation.error ?? "unknown restore error", opts.throwOnError);
     }
   } else {
     // 2b. Lock config file to read-only.
-    //     Uses direct privileged exec to bypass Landlock (same as shields down).
+    //     Uses kubectl exec to bypass Landlock (same as shields down).
     //     Each operation runs independently and the result is verified.
     //     If verification fails, config remains unlocked — we do not lie about state.
     const target = resolveAgentConfig(sandboxName);
@@ -1120,7 +1175,7 @@ function shieldsUp(sandboxName: string, opts: { throwOnError?: boolean } = {}): 
         "  Config remains unlocked — manual intervention required.",
       );
       console.error(
-        `  Restore sandbox access, then run: nemoclaw ${sandboxName} shields up`,
+        `  Re-lock manually via kubectl exec, then run: nemoclaw ${sandboxName} shields up`,
       );
       return failShieldsCommand(message, opts.throwOnError);
     }
@@ -1209,14 +1264,6 @@ function shieldsStatus(
       // Cross-check the sandbox filesystem so a host-root tamper that reverts
       // protected perms back to a sandbox-writable state is surfaced as drift
       // instead of reported as a clean lockdown.
-      //
-      // Scope: DAC only (mode + ownership + legacy-state-layout). The
-      // immutable bit is intentionally out of scope here because shields
-      // state does not persist whether `chattr +i` succeeded at lock time
-      // (`chattr` is best-effort because kubectl exec may lack
-      // CAP_LINUX_IMMUTABLE), so a missing `i` flag at status time cannot
-      // be distinguished from "never set in the first place". Verifying
-      // it requires persisted lock metadata.
       let driftIssues: string[] = [];
       try {
         const target = resolveConfig(sandboxName);
